requirements.txt
Flask==3.1.0
requests==2.32.3
python-dotenv==1.1.0
google-generativeai==0.8.5
langgraph>=0.2.0
langchain-core>=0.3.0
